requirements.txt
mysql-connector-python==8.2.0
b2sdk==1.24.1
schedule==1.2.0
croniter==2.0.1
requests==2.31.0
